# vulkan: scale caching for k quants + misc fixes