--- a/ggml/src/ggml-vulkan/vulkan-shaders/mul_mat_vec_q2_k.comp
+++ b/ggml/src/ggml-vulkan/vulkan-shaders/mul_mat_vec_q2_k.comp
@@ -40,7 +40,7 @@ void compute_outputs(const uint32_t first_row, const uint32_t num_rows) {
 
         [[unroll]] for (uint n = 0; n < num_rows; ++n) {
             const uint ib0 = a_offset / QUANT_K + (first_row+n)*num_blocks_per_row;
-            f16vec2 d = data_a[ib0 + i].d;
+            const f16vec2 d = data_a[ib0 + i].d;
             const FLOAT_TYPE dall = d.x;
             const FLOAT_TYPE dmin = d.y;
 
@@ -48,10 +48,11 @@ void compute_outputs(const uint32_t first_row, const uint32_t num_rows) {
             sccache[ix][1][itid] = FLOAT_TYPE(bitfieldExtract(uint(data_a[ib0 + i].scales[itid8+8]), int(v_im*4), 4)); // upper 8 bytes
             barrier();
 
-            uint16_t qs0_u16 = data_a_packed16[ib0 + i].qs[q_offset / 2];
-            uint16_t qs16_u16 = data_a_packed16[ib0 + i].qs[q_offset / 2 + 8];
-            uvec2 qs0 =  uvec2(unpack8(qs0_u16));
-            uvec2 qs16 = uvec2(unpack8(qs16_u16));
+            const uint32_t qs_u32 = uint32_t(data_a_packed16[ib0 + i].qs[q_offset / 2]) | (uint32_t(data_a_packed16[ib0 + i].qs[q_offset / 2 + 8]) << 16);
+            const vec4 qs_u32_0 = vec4(unpack8(qs_u32 & 0x03030303));
+            const vec4 qs_u32_2 = vec4(unpack8((qs_u32 >> 2) & 0x03030303));
+            const vec4 qs_u32_4 = vec4(unpack8((qs_u32 >> 4) & 0x03030303));
+            const vec4 qs_u32_6 = vec4(unpack8((qs_u32 >> 6) & 0x03030303));
 
             [[unroll]] for (uint j = 0; j < NUM_COLS; ++j) {
                 B_TYPE_VEC2 b0 = data_b_v2[(j*p.batch_stride_b + b_offset + y_idx) / 2];
@@ -66,14 +67,14 @@ void compute_outputs(const uint32_t first_row, const uint32_t num_rows) {
                 FLOAT_TYPE sum1 = FLOAT_TYPE(0.0);
                 FLOAT_TYPE sum2 = FLOAT_TYPE(0.0);
                 [[unroll]] for (int l = 0; l < 2; ++l) {
-                    sum1 = fma(FLOAT_TYPE(b0[l]),   sccache[ix][v_im][0] * FLOAT_TYPE((qs0[l]      ) & 3),
-                           fma(FLOAT_TYPE(b16[l]),  sccache[ix][v_im][1] * FLOAT_TYPE((qs16[l]     ) & 3),
-                           fma(FLOAT_TYPE(b32[l]),  sccache[ix][v_im][2] * FLOAT_TYPE((qs0[l]  >> 2) & 3),
-                           fma(FLOAT_TYPE(b48[l]),  sccache[ix][v_im][3] * FLOAT_TYPE((qs16[l] >> 2) & 3),
-                           fma(FLOAT_TYPE(b64[l]),  sccache[ix][v_im][4] * FLOAT_TYPE((qs0[l]  >> 4) & 3),
-                           fma(FLOAT_TYPE(b80[l]),  sccache[ix][v_im][5] * FLOAT_TYPE((qs16[l] >> 4) & 3),
-                           fma(FLOAT_TYPE(b96[l]),  sccache[ix][v_im][6] * FLOAT_TYPE((qs0[l]  >> 6) & 3),
-                           fma(FLOAT_TYPE(b112[l]), sccache[ix][v_im][7] * FLOAT_TYPE((qs16[l] >> 6) & 3), sum1))))))));
+                    sum1 = fma(FLOAT_TYPE(b0[l]),   sccache[ix][v_im][0] * qs_u32_0[l  ],
+                           fma(FLOAT_TYPE(b16[l]),  sccache[ix][v_im][1] * qs_u32_0[l+2],
+                           fma(FLOAT_TYPE(b32[l]),  sccache[ix][v_im][2] * qs_u32_2[l  ],
+                           fma(FLOAT_TYPE(b48[l]),  sccache[ix][v_im][3] * qs_u32_2[l+2],
+                           fma(FLOAT_TYPE(b64[l]),  sccache[ix][v_im][4] * qs_u32_4[l  ],
+                           fma(FLOAT_TYPE(b80[l]),  sccache[ix][v_im][5] * qs_u32_4[l+2],
+                           fma(FLOAT_TYPE(b96[l]),  sccache[ix][v_im][6] * qs_u32_6[l  ],
+                           fma(FLOAT_TYPE(b112[l]), sccache[ix][v_im][7] * qs_u32_6[l+2], sum1))))))));
                     sum2 = fma(FLOAT_TYPE(b0[l]),   sccache[ix][v_im][ 8],
                            fma(FLOAT_TYPE(b16[l]),  sccache[ix][v_im][ 9],
                            fma(FLOAT_TYPE(b32[l]),  sccache[ix][v_im][10],